+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Simplify `Application::invalidSessionRedirect`
 * Add Abstract Class `Controller`
 * Trait: `CallMethod`
+* Helpers: `loadenv`, `env`
 
 ## [v0.0.02](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.01...v0.0.02) - 8/26/2024